+++ src/main/java/com/pds/project/Models/Vehiculo.java
@@ -18,7 +18,7 @@ import jakarta.persistence.Table;
 public class Vehiculo {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Integer idVehiculo;
+    private Long idVehiculo;
 
     @Column(name = "marca", length = 50)
     private String marca;
